// src/types/api/service/models/NameCountDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MediaBee Service API
 * MediaBee Service API
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface NameCountDto
 */
export interface NameCountDto {
    /**
     * 
     * @type {string}
     * @memberof NameCountDto
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof NameCountDto
     */
    count?: number;
}

/**
 * Check if a given object implements the NameCountDto interface.
 */
export function instanceOfNameCountDto(value: object): value is NameCountDto {
    return true;
}

export function NameCountDtoFromJSON(json: any): NameCountDto {
    return NameCountDtoFromJSONTyped(json, false);
}

export function NameCountDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): NameCountDto {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'count': json['count'] == null ? undefined : json['count'],
    };
}

export function NameCountDtoToJSON(value?: NameCountDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'name': value['name'],
        'count': value['count'],
    };
}